Created a RequesterClass for Connections
This was due to me wanting to keep the main Beepsend class as clean as
possible. Much of the code will be duplicated and I feel that it is more
clean to do it in its own classes. Plus that it makes it more easy (for
me) to locate functions.